+++ apps/backend/src/shared/constants.ts
@@ -4,11 +4,4 @@ export const BANDWAGON_SERVICE_INFO_URL = `${BANDWAGON_URL}/getLiveServiceInfo`
 
 export const BANDWAGON_USAGE_STATS_URL = `${BANDWAGON_URL}/getRawUsageStats`
 
-export const GOOGLE_RECAPTCHA_URL =
-  'https://www.google.com/recaptcha/api/siteverify'
-
-export const IP_STACK_URL = 'http://api.ipstack.com/'
-
 export const SCHEMA_GQL_FILE_NAME = 'schema.gql'
-
-export const TOTP_ENCODE = 'base32'