+++ apps/cms/src/containers/Dashboard/components/TagClouds.tsx
@@ -11,6 +11,9 @@ interface Props {
 const useStyles = makeStyles(() =>
   createStyles({
     paper: {
+      display: 'flex',
+      flex: 2,
+      flexDirection: 'column',
       padding: 16,
       overflowY: 'scroll',
       boxShadow:
@@ -39,16 +42,17 @@ const TagClouds: FC<Props> = ({ tags, loading }) => {
         <PostRankListSkeleton />
       ) : (
         <Paper className={classes.paper}>
-          {' '}
           <header className={classes.header}>Tag Clouds</header>
-          {tags.map((tag) => (
-            <Chip
-              key={tag}
-              label={tag}
-              color="primary"
-              className={classes.chip}
-            />
-          ))}
+          <div>
+            {tags.map((tag) => (
+              <Chip
+                key={tag}
+                label={tag}
+                color="primary"
+                className={classes.chip}
+              />
+            ))}
+          </div>
         </Paper>
       )}
     </>