+++ apps/cms/src/containers/Dashboard/styles.ts
@@ -2,17 +2,31 @@ import { createStyles, makeStyles } from '@mui/styles'
 
 const useStyles = makeStyles(() =>
   createStyles({
-    dashboradWrapper: {
-      width: '100%'
+    dashboardWrapper: {
+      width: '100%',
+      display: 'flex',
+      flexDirection: 'column',
+      gap: 16
     },
 
     group: {
-      display: 'grid',
-      gridTemplateColumns: '2fr 1fr',
-      gridColumnGap: 24,
-      gridTemplateRows: '375px 375px 375px',
-      gridRowGap: 24,
-      marginBottom: 24
+      display: 'flex',
+      gap: 16
+    },
+
+    serverStatusWrapper: {
+      width: '100%',
+      display: 'flex',
+      flex: 1,
+      flexDirection: 'column',
+      boxShadow:
+        'rgb(145 158 171 / 24%) 0px 0px 2px 0px, rgb(145 158 171 / 24%) 0px 16px 32px -4px!important',
+      borderRadius: '16px!important'
+    },
+
+    serverStatusItem: {
+      borderRadius: 16,
+      border: 0
     }
   })
 )