+++ apps/cms/src/layouts/components/Main/styles.ts
@@ -4,7 +4,7 @@ const useStyles = makeStyles({
   main: {
     display: 'flex',
     flex: 1,
-    padding: '72px 16px 72px'
+    padding: '72px 16px 69px'
   }
 })