+++ apps/backend/CHANGELOG.md
@@ -1,5 +1,17 @@
 # Changelog
 
+## 3.1.4
+
+### Patch Changes
+
+- fix cors
+
+## 3.1.3
+
+### Patch Changes
+
+- adapt new config of keycloak
+
 ## 3.1.2
 
 ### Patch Changes